pyproject.toml
[project]
name = "factory-bot-llm-poc"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.76.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "streamlit (>=1.44.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
